moved user guides from static folder to frontend/lib
